#11 Refactoring
 * All tests working except for TranslateTest, which I was working
 on previously.  ie. Not failing due to refactoring.
