+++ src/main/java/au/gov/ga/geodesy/support/dozer/TimePrimitivePropertyTypeEffectiveDatesConverter.java
@@ -1,4 +1,4 @@
-package au.gov.ga.geodesy.domain.converter;
+package au.gov.ga.geodesy.support.dozer;
 
 import java.text.ParseException;
