+++ src/main/java/au/gov/ga/geodesy/support/dozer/TimePrimitivePropertyTypeUtils.java
@@ -1,4 +1,4 @@
-package au.gov.ga.geodesy.domain.converter;
+package au.gov.ga.geodesy.support.dozer;
 
 import java.util.ArrayList;
 import java.util.Date;
