+++ src/test/java/au/gov/ga/geodesy/support/dozer/CodeTypeConverterTest.java
@@ -1,8 +1,9 @@
-package au.gov.ga.geodesy.domain.converter;
+package au.gov.ga.geodesy.support.dozer;
 
 import org.junit.Assert;
 import org.junit.Test;
 
+import au.gov.ga.geodesy.support.dozer.CodeTypeConverter;
 import au.gov.xml.icsm.geodesyml.v_0_2_2.IgsRadomeModelCodeType;
 import net.opengis.gml.v_3_2_1.CodeType;
