+++ src/test/java/au/gov/ga/geodesy/support/dozer/TimePostitionTypeStringConverterTest.java
@@ -1,10 +1,11 @@
-package au.gov.ga.geodesy.domain.converter;
+package au.gov.ga.geodesy.support.dozer;
 
 import java.util.Date;
 
 import org.junit.Assert;
 import org.junit.Test;
 
+import au.gov.ga.geodesy.support.dozer.TimePostitionTypeStringConverter;
 import au.gov.ga.geodesy.support.utils.GMLDateUtils;
 import net.opengis.gml.v_3_2_1.TimePositionType;
